+++ @kiva/kv-components/src/vue/stories/KvCompactLoanCard.stories.js
@@ -410,3 +410,23 @@ export const CustomHref = story({
 	externalLinks: true,
 	customHref: 'https://www.kiva.org',
 });
+
+export const BusinessAndWebsite = story({
+	loanId: loan.id,
+	loan,
+	kvTrackFunction,
+	photoPath,
+	businessName: 'Business Name',
+	website: 'https://www.kiva.org',
+	showLoanUse: false,
+	showTags: false,
+	customHref: '/lend-classic/1?minimal=false',
+});
+
+export const HiddenLoanUse = story({
+	loanId: loan.id,
+	loan,
+	kvTrackFunction,
+	photoPath,
+	showLoanUse: false,
+});